requirements_plus_llm.txt
streamlit
pandas
matplotlib
python-dotenv
PyPDF2
openai
cerebras
transformers
langchain
langchain-community
langchain-text-splitters
sentence-transformers
